refactor(mobile): extract Knowledge Graph WebView HTML builder
KnowledgeGraphWebView.tsx had grown past 1,000 lines, most of it the
buildGraphHtml template literal (bundled D3 + the injected force-sim and
radial-sunburst runtime JS). Move that builder verbatim into
components/graph/graph-webview-html.ts (byte-identical body, only D3_JS as a
dependency), leaving the React component focused on state, WebView messaging
and toolbar chrome. Host drops 1069 -> 660 lines; no behaviour or styling
change.

Verified: mobile tsc + eslint . clean; the extracted string is byte-identical
to the original (diff-confirmed); mobile vitest project (26) green.
