+++ mobile/changelogs/v0.1.3.md
@@ -68,3 +68,4 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - Moved the on-device semantic-search index queries (note + task embeddings) out of the 1,600-line database-queries file into their own module, and factored the shared SQL fragments into one place. Purely internal reorganisation — every existing import keeps working via a re-export; no behaviour change.
 - Split the AI-settings form into smaller pieces (the provider/reasoning segmented buttons, the labelled input field, and the Private-Cloud-Compute usage bar now live in their own files, sharing one style module). Purely internal; the settings screen looks and behaves exactly the same.
 - Split the project detail screen — its note-row, folder-row, tab segment, filter bar and list-row helpers now live in their own files. Purely internal; the project's Overview / Notes / Board tabs look and behave exactly the same.
+- Moved the Knowledge Graph's WebView document builder (the bundled D3 library plus the force-simulation and radial-sunburst rendering code) out of the 1,000-line graph component into its own file. Purely internal; the graph renders and behaves exactly the same.
